CharacterClass now has a JsonUtil
